+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Yii View Change Log
 
-## 11.0.2 under development
+## 11.1.0 under development
 
-- Chg #276: Allow to pass `null` to `ViewInterface` methods `withBasePath()` and `withContext()` (@vjik) 
+- Chg #276: Allow to pass `null` to `ViewInterface` methods `withBasePath()` and `withContext()` (@vjik)
+- New #278: Add `ViewInterface::deepClone()` method that clone object, including state cloning (@vjik)
 
 ## 11.0.1 October 08, 2024